[IFC][Integration][Line clamp] Block children starting with lines exceeding `maximumLineCountForLineClamp` should be fully collapsed. https://bugs.webkit.org/show_bug.cgi?id=255487
Reviewed by Alan Baradlay.

`-webkit-line-clamp` expects the logical height of the content box to be
collapsed to the last visible line.

At `LineLayout::lastLineIndexForContentHeight()`, `visibleLineCountForLineClamp`
of the current layout state may be greater than `maximumLineCountForLineClamp`
when multiple children are being laid out within the same context of line clamp.
In this case, the last line index should be returned as empty in order for the
individual child block to be fully collapsed.

* LayoutTests/fast/text/layout-overflow-hidden-for-line-clamped-content-expected.html: Added.
* LayoutTests/fast/text/layout-overflow-hidden-for-line-clamped-content.html: Added.
* Source/WebCore/layout/integration/inline/LayoutIntegrationLineLayout.cpp:
(WebCore::LayoutIntegration::LineLayout::lastLineIndexForContentHeight const):

Canonical link: https://commits.webkit.org/263079@main