[WGSL] GlobalVariableRewriter should preserve the order of globals https://bugs.webkit.org/show_bug.cgi?id=255595 rdar://108196735
Reviewed by Myles C. Maxfield.

We were using a HashMap for storing the globals keyed by group and binding, which
meant that we were serializing the globals in arbitrary order. Change the code to
use a vector for the second level, so that the binding order can be preserved. We
might still want to sort the entries by binding, but in any case I believe switching
to a vector is a step in the right direction.

* Source/WebGPU/WGSL/GlobalVariableRewriter.cpp:
(WGSL::RewriteGlobalVariables::collectGlobals):
(WGSL::RewriteGlobalVariables::insertStructs):

Canonical link: https://commits.webkit.org/263121@main