fast/history/page-cache-createImageBitmap.html is crashing https://bugs.webkit.org/show_bug.cgi?id=255755 rdar://108347749
Reviewed by NOBODY (OOPS!).

PendingImageBitmap owns a CompletionHandler. It's possible for the PendingImageBitmap to
be destroyed before it has a chance to call the CompletionHandler, which is causing the
crash. This patch simply makes the CompletionHandler optional, sets it to nullopt when
calling it, and adds a destructor that calls the CompletionHandler if it's not nullopt.

* Source/WebCore/html/ImageBitmap.cpp: