REGRESSION (259904@main): window.getSelection() is empty for selection inside textarea https://bugs.webkit.org/show_bug.cgi?id=252281
Reviewed by Wenson Hsieh.

Restore the pre-259904@main behavior of getSelection().toString(), which is to return
the string selected within input and textarea elements. This behavior is consistent with Chrome.

The post-259904@main behavior of not including the selected string within input and textarea
elements are consistent with Firefox but this turned out to be not Web compatible with at least
one Apple internal website.

* LayoutTests/editing/selection/selection-toString-input-expected.txt: Added.
* LayoutTests/editing/selection/selection-toString-input.html: Added.
* LayoutTests/editing/selection/selection-toString-textarea-expected.txt: Added.
* LayoutTests/editing/selection/selection-toString-textarea.html: Added.
* LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/email-set-value-expected.txt: Rebaselined.
* Source/WebCore/page/DOMSelection.cpp:
(WebCore::DOMSelection::toString const):

Canonical link: https://commits.webkit.org/263280@main