Refactor the Canvas post-progressing hash salt https://bugs.webkit.org/show_bug.cgi?id=255994 rdar://problem/108564054
Reviewed by NOBODY (OOPS!).

Currently, this salt is computed for every post-processing call. That's
entirely unnecessary and expensive. The salt is also computed using the
existing implementation of a StringHasher (i.e., SuperFastHash). That may not
be secure for this use case, therefore this patch also moves to using SHA-256.

* Source/WebCore/html/CanvasBase.cpp:
(WebCore::CanvasBase::setNoiseInjectionParameters):
(WebCore::CanvasBase::postProcessPixelBufferResults const):
* Source/WebCore/html/CanvasBase.h:
* Source/WebCore/html/HTMLCanvasElement.cpp:
(WebCore::HTMLCanvasElement::HTMLCanvasElement):