[GTK] MiniBrowser: pressing ESC doesn't exit pointer lock mode https://bugs.webkit.org/show_bug.cgi?id=255864
Reviewed by Adrian Perez de Castro.

This is because we have an accelerator to stop loading when ESC is
pressed, but it's always enabled. The key press is always consumed by
the accel and never passed to the view.

* Tools/MiniBrowser/gtk/BrowserWindow.c:
(webViewIsLoadingChanged): Disable the stop page load action when not loading.
(stopPageLoad): Do not close the search bar in stop page load action.

Canonical link: https://commits.webkit.org/263308@main