[WinCairo][Regression 263118@main] Small viewport does end up in broken screenshots https://bugs.webkit.org/show_bug.cgi?id=255937
Reviewed by NOBODY (OOPS!).

If a small viewport like 50x50 is used, on Windows it does not get rendered
at all on the window (since its outside of the window). Since still a
WebView::onSizeEvent event gets emitted with 50x0, we want to ignore it
and not use it as an actual window size in the drawing area.
Page.snapshotRect breaks due to that

* Source/WebKit/UIProcess/CoordinatedGraphics/DrawingAreaProxyCoordinatedGraphics.cpp:
(WebKit::DrawingAreaProxyCoordinatedGraphics::sendUpdateGeometry):